// src/app/api/model/contract.ts
/**
 * Leasing Service
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * Contact: lease-service@allane.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { Vehicle } from './vehicle';
import { Customer } from './customer';


export interface Contract { 
    id?: number;
    monthlyRate: number;
    vehicle: Vehicle;
    customer: Customer;
}